% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match.R
\name{direct_match}
\alias{direct_match}
\title{Direct Match Species & Genus Binomial or Trinomial names}
\usage{
direct_match(df, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{target_df}{Optional custom target table; if \code{NULL}, uses \code{rWCVPdata::wcvp_names}.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{direct_match}, indicating whether the binomial was successfully matched (TRUE) or not (FALSE)
}
\description{
Tries to directly match Genus + Species | Genus + Species + Rank + Infraspecies to \verb{WCVP data}.
}
