% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{fia}
\alias{fia}
\title{Cleaned Master Tree Species List from FIA}
\format{
A data frame with 2169 rows and 2 variables:
\describe{
\item{Orig.Genus}{Genus name of the species binomial}
\item{Orig.Species}{Specific epithet of the species binomial}
}
}
\usage{
fia
}
\description{
A cleaned dataset containing tree species recorded by the
Forest Inventory and Analysis (FIA) program of the U.S. Forest Service.
This dataset is used in the examples and README of the \code{wcvpmatch}
package. The data was downloaded in November 2022 from the official
webpage of the Forest Inventory and Analysis National Program, available
at the following \href{https://research.fs.usda.gov/products/dataandtools/fia-datamart}{link},
and was originally used during the development of the \code{treemendous}
package. For \code{wcvpmatch}, the variable names have been standardized
to \code{Orig.Genus} and \code{Orig.Species}.
}
\keyword{datasets}
