% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_infraspecies_within_species.R
\name{fuzzy_match_infraspecies_within_species}
\alias{fuzzy_match_infraspecies_within_species}
\title{Fuzzy Match Infraspecific Epithet within Species}
\usage{
fuzzy_match_infraspecies_within_species(
  df,
  target_df = NULL,
  max_dist = 1,
  method = "osa"
)
}
\description{
Fuzzy Match Infraspecific Epithet within Species
}
\keyword{internal}
