% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fuzzy_match_species_within_genus.R
\name{fuzzy_match_species_within_genus}
\alias{fuzzy_match_species_within_genus}
\title{Fuzzy Match Species within Genus}
\usage{
fuzzy_match_species_within_genus(
  df,
  target_df = NULL,
  max_dist = 1,
  method = "osa"
)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{target_df}{Optional custom target table; if \code{NULL}, uses \code{rWCVPdata::wcvp_names}.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{fuzzy_match_species_within_genus}, indicating whether the specific epithet was successfully fuzzy matched within the matched genus (TRUE) or not (FALSE).
}
\description{
Tries to fuzzy match the species epithet within a matched genus against WCVP (\code{rWCVPdata::wcvp_names} by default).
}
\examples{
iucn \%>\%
    dplyr::mutate(Orig.Genus = stringr::str_replace(Orig.Genus, '.{1}$', '')) \%>\%
    dplyr::mutate(Matched.Genus = Orig.Genus) \%>\%
    fuzzy_match_species_within_genus()
}
