% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_match.R
\name{genus_match}
\alias{genus_match}
\title{Match Genus name}
\usage{
genus_match(df, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{target_df}{Optional custom target table; if \code{NULL}, uses \code{rWCVPdata::wcvp_names}.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{genus_match}, indicating whether the genus was successfully matched (TRUE) or not (FALSE)
}
\description{
#' Tries to match the genus name to \verb{WCVP data}.
}
\examples{
iucn \%>\% genus_match()
}
