% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/direct_match_species_within_genus.R
\name{wcvp_direct_match_species_within_genus}
\alias{wcvp_direct_match_species_within_genus}
\title{Direct Match Species within Genus}
\usage{
wcvp_direct_match_species_within_genus(df, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{target_df}{Optional custom target table; if \code{NULL}, uses \code{rWCVPdata::wcvp_names}.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{direct_match_species_within_genus}, indicating whether the specific epithet was successfully matched within the matched genus (TRUE) or not (FALSE).
}
\description{
Tries to directly match the specific epithet within an already matched genus in WCVP.
}
