% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/genus_match.R
\name{wcvp_genus_match}
\alias{wcvp_genus_match}
\title{Match Genus name}
\usage{
wcvp_genus_match(df, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{target_df}{Optional custom target table; if \code{NULL}, uses \code{wcvpdata::wcvp_checklist_names}.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{genus_match}, indicating whether the genus was successfully matched (\code{TRUE}) or not (\code{FALSE}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}

Tries to match the genus name to the 'WCVP' table (\code{wcvpdata::wcvp_checklist_names} by default).
}
\examples{
\dontshow{if (rlang::is_installed("wcvpdata")) withAutoprint(\{ # examplesIf}
library(wcvpmatch)
df <- data.frame(Orig.Genus = "Opuntia", Orig.Species = "yanganucensis")
wcvp_genus_match(df)
\dontshow{\}) # examplesIf}
}
