% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suffix_match_species_within_genus.R
\name{wcvp_suffix_match_species_within_genus}
\alias{wcvp_suffix_match_species_within_genus}
\title{Suffix Match Species within Genus}
\usage{
wcvp_suffix_match_species_within_genus(df, target_df = NULL)
}
\arguments{
\item{df}{\code{tibble} containing the species binomial split into the columns \code{Orig.Genus} and \code{Orig.Species}.}

\item{target_df}{Optional custom target table; if \code{NULL}, uses \code{wcvpdata::wcvp_checklist_names}.}
}
\value{
Returns a \code{tibble} with the additional logical column \code{suffix_match_species_within_genus}, indicating whether the specific epithet was successfully matched within the matched genus (\code{TRUE}) or not (\code{FALSE}).
}
\description{
Tries to match the specific epithet by exchanging common suffixes within an already matched genus in 'WCVP'.
The following suffixes are captured: \code{c("a", "i", "is", "um", "us", "ae")}.
}
\examples{
\dontshow{if (rlang::is_installed("wcvpdata")) withAutoprint(\{ # examplesIf}
library(wcvpmatch)
df <- data.frame(Orig.Genus = "Opuntia", Orig.Species = "yanganucensa", Matched.Genus = "Opuntia")
wcvp_suffix_match_species_within_genus(df)
\dontshow{\}) # examplesIf}
}
